app/requirements.txt
fastapi[all]
uvicorn
pandas
icecream
isatools
pyyaml